Update to latest simplemde, FINALLY.
