+++ WcaOnRails/Gemfile
@@ -74,12 +74,6 @@ gem 'webpacker'
 
 source 'https://rails-assets.org' do
   gem 'rails-assets-autosize'
-
-  # simplemde is pinned to 1.10.1 because version 1.11.0 doesn't give us compiled assets.
-  # See
-  # https://github.com/NextStepWebs/simplemde-markdown-editor/issues/407 and
-  # https://github.com/tenex/rails-assets/issues/367.
-  gem 'rails-assets-simplemde', '1.10.1'
   gem 'rails-assets-autoNumeric'
 end
