+++ WcaOnRails/app/assets/stylesheets/application.css.scss
@@ -28,7 +28,6 @@
 @import "flags/basic";
 @import "flags/flags16";
 @import "flags/flags32";
-@import "simplemde";
 @import "bootstrap-table-parts";
 @import "slick";
 @import "slick-theme";
