+++ WcaOnRails/app/views/layouts/application.html.erb
@@ -11,6 +11,8 @@
   <title><%= full_title(yield(:title)) %></title>
   <%= stylesheet_link_tag 'application', media: 'all' %>
   <%= javascript_include_tag 'application' %>
+
+  <%= stylesheet_pack_tag 'application' %>
   <%= javascript_pack_tag 'application' %>
 
   <% if content_for :include_gmaps %>
